requirements.txt
python-dotenv
requests
colorama
prompt_toolkit
googlesearch-python